+++ src/LudusaviPlayniteSettings.cs
@@ -24,53 +24,64 @@ namespace LudusaviPlaynite
         }
 
         [JsonIgnore]
-        public string BrowseButton_Label { get; set; } = new Translator().BrowseButton();
+        public string BrowseButton_Label { get; set; }
         [JsonIgnore]
-        public string OpenButton_Label { get; set; } = new Translator().OpenButton();
+        public string OpenButton_Label { get; set; }
 
         private string executablePath = "ludusavi";
         public string ExecutablePath { get { return executablePath; } set { executablePath = value; NotifyPropertyChanged("ExecutablePath"); } }
         [JsonIgnore]
-        public string ExecutablePath_Label { get; set; } = new Translator().ExecutablePath_Label();
+        public string ExecutablePath_Label { get; set; }
 
         private string backupPath = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.UserProfile), "ludusavi-playnite");
         public string BackupPath { get { return backupPath; } set { backupPath = value; NotifyPropertyChanged("BackupPath"); } }
         [JsonIgnore]
-        public string BackupPath_Label { get; set; } = new Translator().BackupPath_Label();
+        public string BackupPath_Label { get; set; }
 
         public bool DoBackupOnGameStopped { get; set; } = true;
         [JsonIgnore]
-        public string DoBackupOnGameStopped_Label { get; set; } = new Translator().DoBackupOnGameStopped_Label();
+        public string DoBackupOnGameStopped_Label { get; set; }
 
         public bool AskBackupOnGameStopped { get; set; } = true;
         [JsonIgnore]
-        public string AskBackupOnGameStopped_Label { get; set; } = new Translator().AskBackupOnGameStopped_Label();
+        public string AskBackupOnGameStopped_Label { get; set; }
 
         public bool OnlyBackupOnGameStoppedIfPc { get; set; } = true;
         [JsonIgnore]
-        public string OnlyBackupOnGameStoppedIfPc_Label { get; set; } = new Translator().OnlyBackupOnGameStoppedIfPc_Label();
+        public string OnlyBackupOnGameStoppedIfPc_Label { get; set; }
 
         public bool AddSuffixForNonPcGameNames { get; set; } = false;
         [JsonIgnore]
-        public string AddSuffixForNonPcGameNames_Label { get; set; } = new Translator().AddSuffixForNonPcGameNames_Label();
+        public string AddSuffixForNonPcGameNames_Label { get; set; }
 
         public string SuffixForNonPcGameNames { get; set; } = " (<platform>)";
 
         public bool RetryNonPcGamesWithoutSuffix { get; set; } = false;
         [JsonIgnore]
-        public string RetryNonPcGamesWithoutSuffix_Label { get; set; } = new Translator().RetryNonPcGamesWithoutSuffix_Label();
+        public string RetryNonPcGamesWithoutSuffix_Label { get; set; }
 
         public bool IgnoreBenignNotifications { get; set; } = false;
         [JsonIgnore]
-        public string IgnoreBenignNotifications_Label { get; set; } = new Translator().IgnoreBenignNotifications_Label();
+        public string IgnoreBenignNotifications_Label { get; set; }
 
         // Parameterless constructor must exist if you want to use LoadPluginSettings method.
         public LudusaviPlayniteSettings()
         {
         }
 
-        public LudusaviPlayniteSettings(LudusaviPlaynite plugin)
+        public LudusaviPlayniteSettings(LudusaviPlaynite plugin, Translator translator)
         {
+            BrowseButton_Label = translator.BrowseButton();
+            OpenButton_Label = translator.OpenButton();
+            ExecutablePath_Label = translator.ExecutablePath_Label();
+            BackupPath_Label = translator.BackupPath_Label();
+            DoBackupOnGameStopped_Label = translator.DoBackupOnGameStopped_Label();
+            AskBackupOnGameStopped_Label = translator.AskBackupOnGameStopped_Label();
+            OnlyBackupOnGameStoppedIfPc_Label = translator.OnlyBackupOnGameStoppedIfPc_Label();
+            AddSuffixForNonPcGameNames_Label = translator.AddSuffixForNonPcGameNames_Label();
+            RetryNonPcGamesWithoutSuffix_Label = translator.RetryNonPcGamesWithoutSuffix_Label();
+            IgnoreBenignNotifications_Label = translator.IgnoreBenignNotifications_Label();
+
             // Injecting your plugin instance is required for Save/Load method because Playnite saves data to a location based on what plugin requested the operation.
             this.plugin = plugin;
             Load();